migrated source solution to netstandard
