+++ src/Screenmedia.Plugin.Vanilla.sln
@@ -1,13 +1,15 @@
 
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Screenmedia.Plugin.Vanilla", "Screenmedia.Plugin.Vanilla\Screenmedia.Plugin.Vanilla.csproj", "{5A0581FF-69A3-443B-B7BE-DDE7C9D716A4}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Screenmedia.Plugin.Vanilla.iOS", "Screenmedia.Plugin.Vanilla.iOS\Screenmedia.Plugin.Vanilla.iOS.csproj", "{80ADBC27-60BD-42B4-A7D4-B3FA78EDB3FF}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Screenmedia.Plugin.Vanilla", "Screenmedia.Plugin.Vanilla\Screenmedia.Plugin.Vanilla.csproj", "{29EB9AF0-22C9-4ED4-9239-7262850525CC}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Screenmedia.Plugin.Vanilla.Abstractions", "Screenmedia.Plugin.Vanilla.Abstractions\Screenmedia.Plugin.Vanilla.Abstractions.csproj", "{88E53FD9-91CC-486A-B815-837B7E55132C}"
 EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Screenmedia.Plugin.Vanilla.Shared", "Screenmedia.Plugin.Vanilla.Shared\Screenmedia.Plugin.Vanilla.Shared.shproj", "{F7C954A7-E4AD-4966-B247-9647C80EE350}"
+EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Screenmedia.Plugin.Vanilla.Android", "Screenmedia.Plugin.Vanilla.Android\Screenmedia.Plugin.Vanilla.Android.csproj", "{81D3B0D4-148D-477C-A85F-D1119C2F6934}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Screenmedia.Plugin.Vanilla.iOS", "Screenmedia.Plugin.Vanilla.iOS\Screenmedia.Plugin.Vanilla.iOS.csproj", "{80ADBC27-60BD-42B4-A7D4-B3FA78EDB3FF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,14 +17,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5A0581FF-69A3-443B-B7BE-DDE7C9D716A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5A0581FF-69A3-443B-B7BE-DDE7C9D716A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5A0581FF-69A3-443B-B7BE-DDE7C9D716A4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5A0581FF-69A3-443B-B7BE-DDE7C9D716A4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{80ADBC27-60BD-42B4-A7D4-B3FA78EDB3FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{80ADBC27-60BD-42B4-A7D4-B3FA78EDB3FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{80ADBC27-60BD-42B4-A7D4-B3FA78EDB3FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{80ADBC27-60BD-42B4-A7D4-B3FA78EDB3FF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{29EB9AF0-22C9-4ED4-9239-7262850525CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{29EB9AF0-22C9-4ED4-9239-7262850525CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{29EB9AF0-22C9-4ED4-9239-7262850525CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{29EB9AF0-22C9-4ED4-9239-7262850525CC}.Release|Any CPU.Build.0 = Release|Any CPU
 		{88E53FD9-91CC-486A-B815-837B7E55132C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{88E53FD9-91CC-486A-B815-837B7E55132C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{88E53FD9-91CC-486A-B815-837B7E55132C}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -31,5 +29,9 @@ Global
 		{81D3B0D4-148D-477C-A85F-D1119C2F6934}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{81D3B0D4-148D-477C-A85F-D1119C2F6934}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{81D3B0D4-148D-477C-A85F-D1119C2F6934}.Release|Any CPU.Build.0 = Release|Any CPU
+		{80ADBC27-60BD-42B4-A7D4-B3FA78EDB3FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{80ADBC27-60BD-42B4-A7D4-B3FA78EDB3FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{80ADBC27-60BD-42B4-A7D4-B3FA78EDB3FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{80ADBC27-60BD-42B4-A7D4-B3FA78EDB3FF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
